WIP: Initial work fixing tests
If you add a checkbox to a Rails form the checkbox will always be
submitted as '0', even when it's unchecked (the `check_box` helper method
does this automatically). As a result, any forms containing nested
attributes that have a checkbox on them always pass a '0' for the blank
checkbox.

If no other data is completed on the nested form you might not
want to try and create an object on the nested association. You'd
normally prevent Rails from adding an empty object to the association by
adding `reject_if: :all_blank` on the call to
`accepts_nested_attributes_for`, but that doesn't work in this case
because the check box value is zero, not blank. So we need to specify
our own proc/method. It's already there, but it doesn't know how to
handle unnumbered_command_paper and unnumbered_hoc_paper.

This commit has started to inform it, but as you'll see, the code is
complicated and doesn't work.

I recommend we TDD our way out of here, either by adding tests to
AttachableTest, or by extracting the
`no_substantive_attachment_attributes` proc into an object.

Oh yeah, you'll also want to regenerate the locale translations (the
YAML files).